Start on new event-driven ModuleManager
