+++ library/Zend/Module/Manager.php
@@ -3,9 +3,6 @@
 namespace Zend\Module;
 
 use Traversable,
-    Zend\Config\Config,
-    Zend\Config\Writer\ArrayWriter,
-    Zend\Stdlib\IteratorToArray,
     Zend\EventManager\EventCollection,
     Zend\EventManager\EventManager;
 
@@ -20,26 +17,6 @@ class Manager
      * @var EventCollection
      */
     protected $events;
-
-    /**
-     * @var ManagerOptions
-     */
-    protected $options;
-
-    /**
-     * @var array
-     */
-    protected $mergedConfig = array();
-
-    /**
-     * @var Config
-     */
-    protected $mergedConfigObject;
-
-    /**
-     * @var bool
-     */
-    protected $skipConfig = false;
 
     /**
      * modules 
@@ -59,19 +36,10 @@ class Manager
      * __construct 
      * 
      * @param array|Traversable $modules 
-     * @param ManagerOptions $options 
      * @return void
      */
-    public function __construct($modules, ManagerOptions $options = null)
+    public function __construct($modules)
     {
-        if ($options === null) {
-            $options = new ManagerOptions;
-        }
-        $this->setOptions($options);
-        if ($this->hasCachedConfig()) {
-            $this->skipConfig = true;
-            $this->setMergedConfig($this->getCachedConfig());
-        }
         $this->setModules($modules);
     }
 
@@ -88,23 +56,9 @@ class Manager
         foreach ($this->getModules() as $moduleName) {
             $this->loadModule($moduleName);
         }
-        if ($this->getOptions()->getEnableDependencyCheck()) {
-            $this->resolveDependencies();
-        }
-        $this->updateCache();
         $this->events()->trigger('init.post', $this);
         $this->modulesLoaded = true;
         return $this;
-    }
-
-    /**
-     * Returns boolean representing if modules have been loaded yet 
-     * 
-     * @return Manager
-     */
-    public function modulesLoaded()
-    {
-        return $this->modulesLoaded;
     }
 
     /**
@@ -126,60 +80,10 @@ class Manager
             }
             
             $module = new $class;
-            $this->runModuleInit($module);
-            $this->mergeModuleConfig($module);
+            $this->events()->trigger(__FUNCTION__, $this, array('module' => $module));
             $this->loadedModules[$moduleName] = $module;
         }
         return $this->loadedModules[$moduleName];
-    }
-
-    /**
-     * Set the event manager instance used by this context
-     * 
-     * @param  EventCollection $events 
-     * @return Manager
-     */
-    public function setEventManager(EventCollection $events)
-    {
-        $this->events = $events;
-        return $this;
-    }
-
-    /**
-     * Retrieve the event manager
-     *
-     * Lazy-loads an EventManager instance if none registered.
-     * 
-     * @return EventCollection
-     */
-    public function events()
-    {
-        if (!$this->events instanceof EventCollection) {
-            $this->setEventManager(new EventManager(array(__CLASS__, get_class($this))));
-        }
-        return $this->events;
-    }
-
-    /**
-     * Get options.
-     *
-     * @return ManagerOptions
-     */
-    public function getOptions()
-    {
-        return $this->options;
-    }
- 
-    /**
-     * Set options 
-     * 
-     * @param ManagerOptions $options 
-     * @return Manager
-     */
-    public function setOptions(ManagerOptions $options)
-    {
-        $this->options = $options;
-        return $this;
     }
 
     /**
@@ -196,104 +100,6 @@ class Manager
         return $this->loadedModules;
     }
 
-    /**
-     * getMergedConfig
-     * 
-     * @param array $returnConfigAsObject 
-     * @return mixed
-     */
-    public function getMergedConfig($returnConfigAsObject = true)
-    {
-        if ($returnConfigAsObject === true) {
-            if ($this->mergedConfigObject === null) {
-                $this->mergedConfigObject = new Config($this->mergedConfig);
-            }
-            return $this->mergedConfigObject;
-        } else {
-            return $this->mergedConfig;
-        }
-    }
-
-    /**
-     * setMergedConfig 
-     * 
-     * @param array $config 
-     * @return Manager
-     */
-    public function setMergedConfig(array $config)
-    {
-        $this->mergedConfig = $config;
-        $this->mergedConfigObject = null; // will generate Config instance on-demand if required
-        return $this;
-    }
-
-    /**
-     * mergeModuleConfig 
-     * 
-     * @param mixed $module 
-     * @return Manager
-     */
-    protected function mergeModuleConfig($module)
-    {
-        if ((false === $this->skipConfig)
-            && (is_callable(array($module, 'getConfig')))
-        ) {
-            $config = $module->getConfig($this->getOptions()->getApplicationEnv());
-            if ($config instanceof Traversable) {
-                $config = IteratorToArray::convert($config);
-            }
-            if (!is_array($config)) {
-                throw new \InvalidArgumentException(
-                    sprintf('getConfig() method of %s must be an array, '
-                    . 'implement the \Traversable interface, or be an '
-                    . 'instance of Zend\Config\Config', get_class($module))
-                );
-            }
-            $this->mergedConfig = array_replace_recursive($this->mergedConfig, $config);
-        }
-        return $this;
-    }
-
-    protected function runModuleInit($module)
-    {
-        if (is_callable(array($module, 'init'))) {
-            $module->init($this);
-        }
-        return $this;
-    }
-
-    protected function hasCachedConfig()
-    {
-        if (($this->getOptions()->getEnableConfigCache())
-            && (file_exists($this->getOptions()->getCacheFilePath()))
-        ) {
-            return true;
-        }
-        return false;
-    }
-
-    protected function getCachedConfig()
-    {
-        return include $this->getOptions()->getCacheFilePath();
-    }
-
-    protected function updateCache()
-    {
-        if (($this->getOptions()->getEnableConfigCache())
-            && (false === $this->skipConfig)
-        ) {
-            $this->saveConfigCache($this->getMergedConfig(false));
-        }
-        return $this;
-    }
-
-    protected function saveConfigCache($config)
-    {
-        $content = "<?php\nreturn " . var_export($config, 1) . ';';
-        file_put_contents($this->getOptions()->getCacheFilePath(), $content);
-        return $this;
-    }
- 
     /**
      * Get modules.
      *
@@ -321,5 +127,32 @@ class Manager
             );
         }
         return $this;
+    }
+    
+    /**
+     * Set the event manager instance used by this context
+     * 
+     * @param  EventCollection $events 
+     * @return Manager
+     */
+    public function setEventManager(EventCollection $events)
+    {
+        $this->events = $events;
+        return $this;
+    }
+
+    /**
+     * Retrieve the event manager
+     *
+     * Lazy-loads an EventManager instance if none registered.
+     * 
+     * @return EventCollection
+     */
+    public function events()
+    {
+        if (!$this->events instanceof EventCollection) {
+            $this->setEventManager(new EventManager(array(__CLASS__, get_class($this))));
+        }
+        return $this->events;
     }
 }
