+++ library/Zend/Mvc/Bootstrap.php
@@ -3,10 +3,10 @@ namespace Zend\Mvc;
 
 use Zend\Di\Configuration as DiConfiguration,
     Zend\Di\Di,
+    Zend\Config\Config,
     Zend\EventManager\EventCollection as Events,
     Zend\EventManager\EventManager,
     Zend\EventManager\StaticEventManager,
-    Zend\Module\Manager as ModuleManager,
     Zend\Mvc\Router\Http\TreeRouteStack as Router;
 
 class Bootstrap implements Bootstrapper
@@ -17,11 +17,6 @@ class Bootstrap implements Bootstrapper
     protected $config;
 
     /**
-     * @var ModuleManager
-     */
-    protected $modules;
-
-    /**
      * @var EventCollection
      */
     protected $events;
@@ -29,13 +24,12 @@ class Bootstrap implements Bootstrapper
     /**
      * Constructor
      *
-     * @param  ModuleManager $modules 
+     * @param Config $config 
      * @return void
      */
-    public function __construct(ModuleManager $modules)
+    public function __construct(Config $config)
     {
-        $this->modules = $modules; 
-        $this->setupModules();
+        $this->config = $config; 
     }
 
     /**
@@ -124,7 +118,7 @@ class Bootstrap implements Bootstrapper
     /**
      * Trigger the "bootstrap" event
      *
-     * Triggers with the keys "application" and "modules", the latter pointing
+     * Triggers with the keys "application" and "config", the latter pointing
      * to the Module Manager attached to the bootstrap.
      * 
      * @param  AppContext $application 
@@ -134,20 +128,8 @@ class Bootstrap implements Bootstrapper
     {
         $params = array(
             'application' => $application,
-            'modules'     => $this->modules,
+            'config'      => $this->config,
         );
         $this->events()->trigger('bootstrap', $this, $params);
-    }
-
-    /**
-     * Load modules and pull merged config
-     *
-     * @access protected
-     * @return void
-     */
-    protected function setupModules()
-    {
-        $this->modules->loadModules();
-        $this->config  = $this->modules->getMergedConfig();
     }
 }
